refactor(layout): align collaboration phase ownership
