+++ lib/layout/collaboration/Helpers.js
@@ -7,6 +7,13 @@ export function collectCollaborationShapes(layout) {
   ]);
 }
 
+export function includeHorizontalRange(bounds, min, max) {
+  const right = bounds.x + bounds.width;
+
+  bounds.x = Math.min(bounds.x, min);
+  bounds.width = Math.max(right, max) - bounds.x;
+}
+
 export function collectMessageFlowEndpointDirections(messageFlows) {
   const directions = new Map();
   const addDirection = (element, direction) => {
