+++ lib/layout/collaboration/index.js
@@ -1,10 +1,16 @@
 import { createLayout } from '../geometry/index.js';
+import {
+  orderParticipants
+} from './ordering/ParticipantOrdering.js';
+import {
+  positionParticipants
+} from './placement/ParticipantPlacement.js';
+import {
+  routeMessageFlows
+} from './routing/MessageFlowRouting.js';
 import { compactParticipantRows } from './steps/compactParticipantRows.js';
 import { layoutParticipants } from './steps/layoutParticipants.js';
-import { orderParticipants } from './steps/orderParticipants.js';
 import { placeArtifacts } from './steps/placeArtifacts.js';
-import { positionParticipants } from './steps/positionParticipants.js';
-import { routeMessageFlows } from './steps/routeMessageFlows.js';
 import {
   validateCollaboration
 } from './steps/validateCollaboration.js';
