+++ lib/layout/collaboration/placement/ParticipantPlacement.js
@@ -10,18 +10,90 @@ import {
   point,
   bounds,
   compareScores,
+  getExpandedChildEdges,
   routeLength,
+  translateLayout,
   toSegments,
   segmentsProperlyCross
 } from '../../geometry/index.js';
 import {
   getMessageObstacles,
-  routeMessageFlows
+  routeAllMessageFlows
 } from '../routing/MessageFlowRouting.js';
 import {
   findEndpointParticipant,
   resolveMessageFlowEndpoint
 } from '../EndpointResolution.js';
+import {
+  collectCollaborationShapes,
+  includeHorizontalRange
+} from '../Helpers.js';
+
+/**
+ * @typedef {import('../../Types.js').CollaborationLayoutContext} CollaborationLayoutContext
+ */
+
+/**
+ * @param {CollaborationLayoutContext} context
+ * @returns {CollaborationLayoutContext}
+ */
+export function positionParticipants(context) {
+  const { collaboration, layout } = context;
+  const {
+    anchorPositioned,
+    expandable,
+    layouts
+  } = context.participants;
+  const { channelOffsets } = context.routing;
+  let shapes = collectCollaborationShapes(layout);
+  const positions = alignParticipantsHorizontally(
+    collaboration,
+    layout.shapes,
+    shapes,
+    getExpandedChildEdges(layout),
+    channelOffsets,
+    anchorPositioned,
+    expandable
+  );
+
+  for (const [ participant, x ] of positions) {
+    const participantBounds = layout.shapes.get(participant);
+    const processLayout = layouts.get(participant);
+    const dx = x - participantBounds.x;
+
+    participantBounds.x = x;
+
+    if (processLayout) {
+      translateLayout(processLayout, dx, 0);
+    }
+  }
+
+  shapes = collectCollaborationShapes(layout);
+  sizeAndPositionParticipantsFromMessageAnchors(
+    collaboration,
+    layout.shapes,
+    shapes,
+    channelOffsets,
+    anchorPositioned,
+    expandable
+  );
+
+  for (const [ participant, dx ] of alignParticipantComponentsLeft(
+    collaboration,
+    layout.shapes
+  )) {
+    const participantBounds = layout.shapes.get(participant);
+    const processLayout = layouts.get(participant);
+
+    participantBounds.x += dx;
+
+    if (processLayout) {
+      translateLayout(processLayout, dx, 0);
+    }
+  }
+
+  return context;
+}
 
 export function alignParticipantsHorizontally(
     collaboration,
@@ -402,7 +474,7 @@ function horizontalAlignmentScore(
   );
 
   const obstacles = getMessageObstacles(translatedShapes);
-  const routes = routeMessageFlows(
+  const routes = routeAllMessageFlows(
     collaboration,
     translatedParticipants,
     translatedShapes,
@@ -644,47 +716,6 @@ function ensureParticipantConnectionOverlap(
   }
 }
 
-export function includeResizableParticipantMessageDocks(
-    collaboration,
-    participantShapes,
-    edges,
-    expandableParticipants) {
-  let changed = false;
-
-  for (const messageFlow of collaboration.messageFlows || []) {
-    const points = edges.get(messageFlow);
-
-    if (!points?.length) {
-      continue;
-    }
-
-    for (const [ endpoint, dock ] of [
-      [ messageFlow.sourceRef, points[0] ],
-      [ messageFlow.targetRef, points.at(-1) ]
-    ]) {
-      if (!is(endpoint, 'bpmn:Participant') || !expandableParticipants.has(endpoint)) {
-        continue;
-      }
-
-      const participantBounds = participantShapes.get(endpoint);
-      const previousX = participantBounds.x;
-      const previousWidth = participantBounds.width;
-
-      includeHorizontalRange(
-        participantBounds,
-        dock.x - PARTICIPANT_HEADER_WIDTH,
-        dock.x + PARTICIPANT_HEADER_WIDTH
-      );
-
-      changed = changed ||
-        participantBounds.x !== previousX ||
-        participantBounds.width !== previousWidth;
-    }
-  }
-
-  return changed;
-}
-
 function ensureParticipantDockOverlap(
     source,
     target,
@@ -727,11 +758,4 @@ function ensureParticipantDockOverlap(
   if (targetResizable) {
     includeHorizontalRange(target, dockStart, dockEnd);
   }
-}
-
-function includeHorizontalRange(bounds, min, max) {
-  const right = bounds.x + bounds.width;
-
-  bounds.x = Math.min(bounds.x, min);
-  bounds.width = Math.max(right, max) - bounds.x;
 }
