+++ lib/layout/collaboration/routing/MessageFlowRouting.js
@@ -21,6 +21,50 @@ import {
   findEndpointParticipant,
   resolveMessageFlowEndpoint
 } from '../EndpointResolution.js';
+import {
+  collectCollaborationShapes,
+  includeHorizontalRange
+} from '../Helpers.js';
+
+/**
+ * @typedef {import('../../Types.js').CollaborationLayoutContext} CollaborationLayoutContext
+ */
+
+/**
+ * @param {CollaborationLayoutContext} context
+ * @returns {CollaborationLayoutContext}
+ */
+export function routeMessageFlows(context) {
+  const { collaboration, layout } = context;
+  const { expandable } = context.participants;
+  const { channelOffsets } = context.routing;
+  const shapes = collectCollaborationShapes(layout);
+  const obstacles = getMessageObstacles(shapes);
+  let participantBoundsChanged;
+
+  do {
+    const routes = routeAllMessageFlows(
+      collaboration,
+      layout.shapes,
+      shapes,
+      obstacles,
+      channelOffsets
+    );
+
+    for (const [ messageFlow, points ] of routes) {
+      layout.edges.set(messageFlow, points);
+    }
+
+    participantBoundsChanged = includeResizableParticipantMessageDocks(
+      collaboration,
+      layout.shapes,
+      layout.edges,
+      expandable
+    );
+  } while (participantBoundsChanged);
+
+  return context;
+}
 
 export function routeMessageFlow(
     source,
@@ -405,7 +449,7 @@ export function getMessageObstacles(shapes) {
     .map(([ element, rect ]) => ({ element, rect }));
 }
 
-export function routeMessageFlows(
+export function routeAllMessageFlows(
     collaboration,
     participantShapes,
     endpointShapes,
@@ -441,6 +485,50 @@ export function routeMessageFlows(
   }
 
   return routes;
+}
+
+function includeResizableParticipantMessageDocks(
+    collaboration,
+    participantShapes,
+    edges,
+    expandableParticipants) {
+  let changed = false;
+
+  for (const messageFlow of collaboration.messageFlows || []) {
+    const points = edges.get(messageFlow);
+
+    if (!points?.length) {
+      continue;
+    }
+
+    for (const [ endpoint, dock ] of [
+      [ messageFlow.sourceRef, points[0] ],
+      [ messageFlow.targetRef, points.at(-1) ]
+    ]) {
+      if (
+        !is(endpoint, 'bpmn:Participant') ||
+        !expandableParticipants.has(endpoint)
+      ) {
+        continue;
+      }
+
+      const participantBounds = participantShapes.get(endpoint);
+      const previousX = participantBounds.x;
+      const previousWidth = participantBounds.width;
+
+      includeHorizontalRange(
+        participantBounds,
+        dock.x - PARTICIPANT_HEADER_WIDTH,
+        dock.x + PARTICIPANT_HEADER_WIDTH
+      );
+
+      changed = changed ||
+        participantBounds.x !== previousX ||
+        participantBounds.width !== previousWidth;
+    }
+  }
+
+  return changed;
 }
 
 function findMessageFlowDockX(
